ログイン周り
未検証　あとでやる
あとRust側の変更検知・再起動をcargo-watchからwatchexecに変更　cargo-watchはメンテ終了してたらしい

+++ app/ondyst/src/utils/mod.rs
@@ -5,7 +5,7 @@ mod state;
 pub use page::{ActorData, Page};
 pub use state::State;
 
-pub type Identity = common::Identity<i32>;
+pub type Identity = common::Identity<i64>;
 pub type StateHandle = common::StateHandle<State>;
 
 pub fn app(path: &str) -> String {
